+++ Changelog.md
@@ -10,6 +10,7 @@ et le projet suit un schéma de versionning inspiré de [Calendar Versioning](ht
 #### :rocket: Nouvelles fonctionnalités
 
 - ETQ utilisateur je peux annuler un BSDA (via le processus de révision) [PR 2200](https://github.com/MTES-MCT/trackdechets/pull/2200)
+- ETQ destinataire d'un BSFF, je peux modifier le numéro de contenant lors de l'acceptation [PR 2208](https://github.com/MTES-MCT/trackdechets/pull/2208)
 
 #### :bug: Corrections de bugs